requirements.txt
croniter==0.3.30
Django==2.2.5
django_compressor==2.2
psycopg2==2.7.5
pytz==2019.1
requests==2.22.0
django-allauth==0.40.0